Remove maximum-scale=1 from viewport meta tag
The maximum-scale=1 attribute prevents low-vision users from magnifying the page in some browsers. This isn't applicable to browsers like Safari in iOS 10+, but it would still be ideal to accommodate the aforementioned users for other browsers. Also, this results in a Lighthouse score penalty for accessibility. Alternatively, we could increase maximum-scale to 5 or above if removing it results in a known UI/UX regression I haven't looked into.

Some further reading:
* https://a11yproject.com/posts/never-use-maximum-scale/
* https://web.dev/meta-viewport
* https://www.w3.org/TR/WCAG21/#resize-text